requirements.txt
# Core dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
scipy>=1.10.0
numpy>=1.24.0

# Image processing
scikit-image>=0.21.0

# Deep learning (optional - for SAM 2)
torch>=2.0.0
torchvision>=0.15.0

# SAM 2 (optional - install separately)
# git+https://github.com/facebookresearch/segment-anything-2.git

# Utilities
requests>=2.31.0
Pillow>=10.0.0

# Optional dependencies for training
# matplotlib>=3.7.0
# tqdm>=4.65.0
